requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==23.1
geopy==2.3.0
python-dotenv # ==1.0.0
Pillow==9.0.0